Add initial engine sound
Currently there are some issues like if the wheels are in the air, even
though you'd expect them to be turning they aren't so if you're applying
throttle in the air it sounds like the engine is decelerating.

The sounds are heard from the perspective of the first player camera,
which could be annoying for other players but I think it will be less
noticable when there's more action and noises going on.

This engine noise I've just taken from stuntrally, I'm waiting to get my
account on freesound.org working so I can get a better one.

There is also no gear change noise, which would add a bit more realism.